% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_analysis.R
\name{compute_indiv_dm_table}
\alias{compute_indiv_dm_table}
\title{Compute differential mehylation table for each probe according to individual-level analysis.}
\usage{
compute_indiv_dm_table(gene_list, exp_grp, data_meth,
  filter_indiv = "no_filter", indiv_filtering_matrix, pf_meth, pf_pos_colname,
  pf_chr_colname, updwn_str = 5000, slide = 0, apply_func = apply)
}
\arguments{
\item{gene_list}{A \code{gene_list} bedfile containing the genes to screen for differential methylation.}

\item{exp_grp}{A \code{exp_grp} dataframe that contains metadatas on individuals and samples.}

\item{data_meth}{A \code{meth} matrix that contains methylation information (beta values). Columns correspond to indivuals, row correspond to probes.}

\item{filter_indiv}{A vector of individual names to be screened for differential methylation. Optionnal (set on "no_filter" by default).}

\item{indiv_filtering_matrix}{A matrix of filter to define reference and tested individual for each probe. Rownames should be contained in rownames(gene_list) and colnames should contain \code{filter_indiv}. Reference individuals should be set to "0" and individuals to be tested should be set to "1".}

\item{pf_meth}{A data frame describing CpG positions.}

\item{pf_pos_colname}{String matching the name of the column in the platform that contain the position information of probes.}

\item{pf_chr_colname}{String matching the name of the column in the platform that contain the chromosome on which we find a probes.}

\item{updwn_str}{An integer specifying up and down stream size (in bp). By default set on 5000pb.}

\item{slide}{The maximum width slide allowed when comparing two curves. By default set on 0.}

\item{apply_func}{A function to be used as/instead of R \code{base::apply}. By default set on \code{base::apply}.}
}
\value{
A matrix of differential methylation values based on single individual analysis.
}
\description{
Generate differential methylation data table from beta values. According to a reference individual matrix which indicates which samples should be considere as reference for a given probe, the function computes the methylation difference between each tested sample and the mean of reference samples.
}
