% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_analysis.R
\name{compute_pop_dm_table}
\alias{compute_pop_dm_table}
\title{Compute differential mehylation table for each probe according to population-level analysis (tested samples VS control).}
\usage{
compute_pop_dm_table(gene_list, exp_grp, data_meth,
  filter_indiv = "no_filter", contrast = c("tissue_status", "patho",
  "normal"), pf_meth, pf_pos_colname, pf_chr_colname, updwn_str = 5000,
  slide = 0, apply_func = apply)
}
\arguments{
\item{gene_list}{A \code{gene_list} bedfile containing the genes to screen for differential methylation.}

\item{exp_grp}{A \code{exp_grp} dataframe that contains metadatas on individuals and samples.}

\item{data_meth}{A \code{meth} matrix that contains methylation information (beta values). Columns correspond to indivuals, row correspond to probes.}

\item{filter_indiv}{A vector of individual names to be screened for differential expression. Optionnal (set on "no_filter" by default).}

\item{contrast}{A vector containing the constrast to use to estimate the differential methylation. By default: c("tissue_status","patho","normal")}

\item{pf_meth}{A data frame describing CpG positions.}

\item{pf_pos_colname}{String matching the name of the column in the platform that contain the position information of probes.}

\item{pf_chr_colname}{String matching the name of the column in the platform that contain the chromosome on which we find a probes.}

\item{updwn_str}{An integer specifying up and down stream size (in bp). By default set on 5000pb.}

\item{slide}{The maximum width slide allowed when comparing two curves. By default set on 0.}

\item{apply_func}{A function to be used as/instead of R \code{base::apply}. By default set on \code{base::apply}.}
}
\value{
A matrix of differential methylation values based on population analysis.
}
\description{
Generate differential methylation data table from beta values. It basically extract all tumorous samples and controls. Then it computes the difference between each tumorous and the mean of control.
}

