% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmDistance_translocate.R
\name{dmDistance_translocate}
\alias{dmDistance_translocate}
\title{dmDistance_translocate}
\usage{
dmDistance_translocate(dmprofileList, win = 5000, slide = 500,
  by.interp = 20)
}
\arguments{
\item{dmprofileList}{a list of dmProfile}

\item{win}{is the width of the window on the chromosome in bp where the function will fetch probes position and differential methylation value, default is 5000.}

\item{slide}{is the maximum width slide you'll alow when comparing two curve, default is 0.}

\item{interp.by}{is resolution at which the function interpolate the probes signal, default is 20.}
}
\description{
Produce a list of two matrix : The distance matrix from a list of dmProfile. Each profile is translocated N times against another, we then keep the min(distance) in the matrix. The second matrix indicates which translocation returned the min(distance)
}
\examples{
### generate random samples ids
randomDataset          <- dmRandomDataset()
### get a methylation differential table
meth_study          <- randomDataset$meth_study
tumoral_ref         <- rownames(meth_study$exp_grp)[meth_study$exp_grp$ref=="case"]
control_ref         <- rownames(meth_study$exp_grp)[meth_study$exp_grp$ref=="ctrl"]
diff_meth_study     <- dmTable(meth_study$data,
                              meth_study$platform,
                              meth_study$exp_grp,
                              tumoral_ref,
                              control_ref)
### get dmProfiles for multiple genes
genes        <- randomDataset$genes 
alldmProfile <- getalldmProfile(diff_meth_study,
                                bedfile = genes, slide = 500,
                                pf_chr_colname = "chr",
                                pf_pos_colname = "pos")

### get a distance matrix from a list of dmProfile while minimizing distance by translocating
dblist <- dmDistance_translocate(dmprofileList = dmprofileList, slide = 500)
}
