% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmProfile.R
\name{dmProfile}
\alias{dmProfile}
\title{dmProfile}
\usage{
dmProfile(gene_study_info, win = 5000, interp.by = 20, slide = 0)
}
\arguments{
\item{gene_study_info}{is a subset of diff_meth_study list for a given gene bedline and a size of window}

\item{win}{is the width of the window on the chromosome in bp where the function will fetch probes position and differential methylation value}

\item{interp.by}{is resolution at which the function interpolate the probes signal}

\item{slide}{is the maximum width slide you'll alow when comparing two curve}
}
\description{
Generate  differential methylation profile for genes around the transcription start side
}
