% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmRandomDataset.R
\name{dmRandomDataset}
\alias{dmRandomDataset}
\title{dmRandomDataset}
\usage{
dmRandomDataset()
}
\description{
Generate random meth_study to run the functions examples.
}
\examples{
### generate random samples ids
meth_study          <- dmRandomDataset()
}
