% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmTable.R
\name{dmTable}
\alias{dmTable}
\title{dmTable}
\usage{
dmTable(data, platform, exp_group, TumoralRef, ControlRef, NARM = TRUE)
}
\arguments{
\item{data}{A matrix of row methylation data with TCGA sample ids as column names and probes ids as row names.}

\item{platform}{a dataframe with metadata types as columns names and probes ids as row names.}

\item{exp_group}{a dataframe of metadata about the samples. Rows are ids and columns information type.}

\item{TumoralRef}{a vector of ids corresponding to tumoral samples.}

\item{ControlRef}{a vector of ids corresponding to control samples.}

\item{NARM}{will delete empty rows in data and platform if TRUE.}
}
\description{
Generate differential methylation data table from beta values. It basically extract all tumorous samples and controls. Then it computes the difference between each tumorous and the mean of control.
}
