% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trscr_analysis.R
\name{fad_RNAseq_diffAnalysis}
\alias{fad_RNAseq_diffAnalysis}
\title{Perform a Quick differential analysis of RNAseq data.}
\usage{
fad_RNAseq_diffAnalysis(data_trscr, exp_grp, gene_list,
  filter_indiv = "no_filter", contrast = c("sample", "01", "11"),
  apply_func = apply)
}
\arguments{
\item{data_trscr}{A \code{data} matrix that contains transcriptome information (RNAseq counts from HTseq). Columns correspond to indivuals, row correspond to genes.}

\item{exp_grp}{A \code{exp_grp} dataframe that contains metadatas on \code{data_trscr} individuals.}

\item{gene_list}{A \code{gene_list} bedfile containing the genes to screen for differential expression.}

\item{filter_indiv}{A vector of individual names to be screened for differential expression. Optionnal (set on "no_filter" by default).}

\item{contrast}{A vector containing the constrast to be used to estimate the logarithmic fols change}

\item{apply_func}{Function to be used for apply.}
}
\value{
A \code{gene_list} table including log2FoldChange and adjusted p-value (padj) computed by DESeq2 and a \code{data_ntrscr} matrix of normalized counts.
}
\description{
Perform a quick differential analysis of transcriptomic data (RNAseq) using non parametric test.
}
