% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_analysis.R
\name{from_big_to_profile}
\alias{from_big_to_profile}
\title{from_big_to_profile}
\usage{
from_big_to_profile(big, xf, mask)
}
\arguments{
\item{big}{matrix of interpolated methylome signal of samples}

\item{xf}{coordinates of interpolation}

\item{mask}{is mask of o probe}
}
\description{
Transform a matrix of interpolated methylome signal of samples to a methyl;ome profile
}
