% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_analysis.R
\name{get_probe_names}
\alias{get_probe_names}
\title{get_probe_names}
\usage{
get_probe_names(gene, pf_meth, pf_chr_colname = "Chromosome",
  pf_pos_colname = "Start", up_str = 5000, dwn_str = 5000)
}
\arguments{
\item{gene}{A vector describing the gene (line of a bed file).}

\item{pf_meth}{A data frame describing CpG positions.}

\item{pf_chr_colname}{string matching the name of the column in the platform that contain the chromosome on which we find a probes.}

\item{pf_pos_colname}{string matching the name of the column in the platform that contain the position information of probes.}

\item{up_str}{An integer specifying up stream size (in bp).}

\item{dwn_str}{An integer specifying down stream size (in bp).}
}
\value{
A vector of probe names
}
\description{
This function extracts probe names of a given gene from platform
}

