% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getalldmProfile.R
\name{getalldmProfile}
\alias{getalldmProfile}
\title{getalldmProfile}
\usage{
getalldmProfile(diff_meth_study, bedfile, nbProbes = 0, win = 5000,
  slide = 0, interp.by = 20, pf_chr_colname = "Chromosome",
  pf_pos_colname = "Start")
}
\arguments{
\item{diff_meth_study}{list composed of a methylation differential table, an exp_grp dataframe and a platform.}

\item{bedfile}{a bedfile contaning information for one gene on each line.}

\item{nbProbes}{will make the function return the profile of genes for which it found more than nbProbes in the window around tss.}

\item{win}{is the width of the window on the chromosome in bp where the function will fetch probes position and differential methylation value.}

\item{slide}{is the maximum width slide you'll alow when comparing two curve.}

\item{interp.by}{is resolution at which the function interpolate the probes signal.}

\item{pf_chr_colname}{string matching the name of the column in the platform that contain the chromosome on which we find a probes.}

\item{pf_pos_colname}{string matching the name of the column in the platform that contain the position information of probes.}
}
\description{
This function is combination of getMethylInfo and dmProfile, it return a list of differential methylation profile. The bedline parameter is now a dataframe of bedline for which each will produce a dmProfile
}
\examples{
### generate random samples ids
randomDataset          <- dmRandomDataset()
### get a methylation differential table
meth_study          <- randomDataset$meth_study
tumoral_ref         <- rownames(meth_study$exp_grp)[meth_study$exp_grp$ref=="case"]
control_ref         <- rownames(meth_study$exp_grp)[meth_study$exp_grp$ref=="ctrl"]
diff_meth_study     <- dmTable(meth_study$data,
                              meth_study$platform,
                              meth_study$exp_grp,
                              tumoral_ref,
                              control_ref)
}
