% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDifferentialTable.R
\name{getDifferentialTable}
\alias{getDifferentialTable}
\title{getDifferentialTable}
\usage{
getDifferentialTable(data, platform, exp_group)
}
\arguments{
\item{data}{A matrix of row methylation data with TCGA sample ids as column names and probes ids as row names}

\item{platform}{a dataframe with metadata types as columns names and probes ids as row names}

\item{exp_group}{a dataframe of metadata about the samples. Rows are ids and columns information type.}
}
\description{
Generate differential methylation data table from beta values. It basically extract all tumorous samples and controls. Then it computes the difference between each tumorous and the mean of control
}
\examples{
library(gsubfn)
###generate random samples ids  
regexTum <- "^TCGA-.{2}-.{4}-01.{1}-.{3}-.{4}-.{2}$"
regexWT <- "^TCGA-.{2}-.{4}-11.{1}-.{3}-.{4}-.{2}$"

sampleId <- c(
         sapply(1:6, function(x) {
             gsub("\\\\^|\\\\$", "",
             gsubfn("\\\\.{([[:digit:]]+)}",
             ~ paste(rep(sample(c(LETTERS, 1:9), 1), n), collapse=""), regexTum)) }),
         sapply(1:4, function(x) {
             gsub("\\\\^|\\\\$", "",
             gsubfn("\\\\.{([[:digit:]]+)}", 
             ~ paste(rep(sample(c(LETTERS, 1:9), 1), n), collapse=""), regexWT)) })
             )
idProbes <- paste0(rep("probe", 10), 1:10) 
data <- matrix(runif(100,0,1), 10, 10)
colnames(data) <- sampleId
rownames(data) <- idProbes

###generate exp_group
exp_group <- as.data.frame(matrix(rep(1, 50),  10, 5), row.names = sampleId) 

###generate platfrom
platform  <-as.data.frame(matrix(rep(1, 50),  10, 5), row.names = idProbes)
diff_table <- getDifferentialTable(data, platform, exp_group)
}
