% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethylInfo.R
\name{getMethylInfo}
\alias{getMethylInfo}
\title{getMethylInfo}
\usage{
getMethylInfo(diff_meth_study, bedline, win = 5000)
}
\arguments{
\item{diff_meth_study}{list composed of a methylation differential table, an exp_grp dataframe and a platform}

\item{bedline}{is a line of of the bedfile dataframe}

\item{win}{is the width of the window on the chromosome in bp where the function will fetch probes position and differential methylation value. Default is 5000 bp}
}
\description{
Return a subset of diff_meth_study list for a given gene bedline and a size of window
}
