% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProfileListID.R
\name{getProfileListID}
\alias{getProfileListID}
\title{getProfileListID}
\usage{
getProfileListID(dmprofileList)
}
\arguments{
\item{dmprofileList}{a list of dmProfile}
}
\value{
This function return a vector of gene names.
}
\description{
Get a vector of Gene names from a list of dmProfile
}
