% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmProfile.R
\name{interpolate_gene}
\alias{interpolate_gene}
\title{interpolate_gene}
\usage{
interpolate_gene(vec, probes_pos, tss, strand, win, slide, interp.by)
}
\arguments{
\item{vec}{A numeric vector specifying differential methylation signal.}

\item{probes_pos}{is a vector of probes position on the chromosome.}

\item{tss}{the transcription start site on the same chromosome.}

\item{strand}{is the strand from which the gene is red.}

\item{win}{is the width of the window on the chromosome in bp where the function will fetch probes position and differential methylation value, default is 5000.}

\item{slide}{is the maximum width slide you'll alow when comparing two curve, default is 0.}

\item{interp.by}{is resolution at which the function interpolate the probes signal, default is 20.}
}
\description{
Build a interpolated signal of differential methylation value for a gene.
}

