% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateGene.R
\name{interpolateGene}
\alias{interpolateGene}
\title{interpolateGene}
\usage{
interpolateGene(n, diff_table, probes_start, promoterPos, strand, win = 5000,
  slide = 0, interp.by = 20)
}
\arguments{
\item{n}{is a vector of index of the column of the data matrix (each corresponding to one sample).}

\item{diff_table}{is a differential methylation data matrix.}

\item{probes_start}{is a vector of probes position on the chromosome.}

\item{promoterPos}{the transcription start site on the same chromosome.}

\item{strand}{is the strand from which the gene is red.}

\item{win}{is the width of the window on the chromosome in bp where the function will fetch probes position and differential methylation value, default is 5000.}

\item{slide}{is the maximum width slide you'll alow when comparing two curve, default is 0.}

\item{interp.by}{is resolution at which the function interpolate the probes signal, default is 20.}
}
\description{
Build a curve from differential methylation value for one sample.
}
