% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_analysis.R
\name{plot_meth_profile}
\alias{plot_meth_profile}
\title{plot_meth_profile}
\usage{
plot_meth_profile(meth_profile, alpha.f, ...)
}
\arguments{
\item{meth_profile}{a list of dmProfile}

\item{alpha.f}{a numeric specifying transparency of convolved signal.}

\item{...}{args pass to plot}
}
\description{
Plot methylome profile for a gene.
}

