% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trscr_meth_analysis.R
\name{plot_trscr_meth_analysis}
\alias{plot_trscr_meth_analysis}
\title{Plot transcriptome and methylome co-analyse}
\usage{
plot_trscr_meth_analysis(results)
}
\arguments{
\item{results}{output of trscr_meth_analysis function}
}
\description{
This plots transcriptome and methylome co-analyse
}

