% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdmProfile.R
\name{plotdmProfile}
\alias{plotdmProfile}
\title{plotdmProfile}
\usage{
plotdmProfile(dmprofileList)
}
\arguments{
\item{dmprofileList}{a list of dmProfile}
}
\description{
Produce a list of differential methylation profile plot with ggplot2.
}
\examples{
### generate random samples ids
randomDataset          <- dmRandomDataset()
### get a methylation differential table
meth_study          <- randomDataset$meth_study
tumoral_ref         <- rownames(meth_study$exp_grp)[meth_study$exp_grp$ref=="case"]
control_ref         <- rownames(meth_study$exp_grp)[meth_study$exp_grp$ref=="ctrl"]
diff_meth_study     <- dmTable(meth_study$data,
                              meth_study$platform,
                              meth_study$exp_grp,
                              tumoral_ref,
                              control_ref)
### get dmProfiles for multiple genes
genes        <- randomDataset$genes 
alldmProfile <- getalldmProfile(diff_meth_study,
                                bedfile = genes, slide = 500,
                                pf_chr_colname = "chr",
                                pf_pos_colname = "pos")

### get a list of plot(s) from a list of profile(s)
somedmProfile <-  alldmProfile[1:10] 
plotlist <- plotdmProfile(somedmProfile)
plotlist[[1]]
}
