% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trscr_analysis.R
\name{RNAseq_normalization}
\alias{RNAseq_normalization}
\title{Normalize RNAseq data according to DESeq2 method.}
\usage{
RNAseq_normalization(data_trscr, exp_grp, gene_list,
  filter_indiv = "no_filter", alpha = 0.05, contrast = c("sample", "01",
  "11"), fitType = "parametric")
}
\arguments{
\item{data_trscr}{A \code{data} matrix that contains transcriptome information (RNAseq counts from HTseq). Columns correspond to indivuals, row correspond to genes.}

\item{exp_grp}{A \code{exp_grp} dataframe that contains metadatas on \code{data_trscr} individuals.}

\item{gene_list}{A \code{gene_list} bedfile containing the genes to screen for differential expression.}

\item{filter_indiv}{A vector of individual names to be screened for differential expression. Optionnal (set on "no_filter" by default).}

\item{alpha}{A parameter to indicate the significance cutoff used by \code{DESeq2::results} funtion for optimizing the independent filtering (by default 0.05). If the adjusted p-value cutoff (FDR) will be a value other than 0.1, alpha should be set to that value.}

\item{contrast}{A vector containing the constrast to be used to estimate the logarithmic fols change}

\item{fitType}{A DESeq fitting paramater, by default set to "parametric"}
}
\value{
A list composed of a \code{data_ntrscr} matrix of normalized counts and a \code{size_factor} vector of normalization factors.
}
\description{
Perform normalization of transcriptomic data (RNAseq) using DESeq2 R package.
}
