% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffAnalysis.R
\name{select_candidates}
\alias{select_candidates}
\title{Generate a candidate table}
\usage{
select_candidates(gene_list, padj_thresh, z_score_thresh, noCNVlog2FC_thresh)
}
\arguments{
\item{gene_list}{A \code{gene_list} bedfile containing the log2FC, zscore and noCNVlog2FC values.}

\item{padj_thresh}{A threshold p-value under which expression is significantly different by DESeq analysis}

\item{z_score_thresh}{A threshold zscore value, candidate genes should display a zcore value such as -t_zscore<zscore<t_zscore}

\item{noCNVlog2FC_thresh}{A threshold noCNVlog2FC, candidate genes should display a noCNVlog2FC value such as noCNVlog2FC > noCNVlog2FC_thresh or noCNVlog2FC < noCNVlog2FC_thresh}
}
\value{
A \code{gene_list} table including of candidate genes
}
\description{
Generate a candidate table
}
\examples{
### generate a list of simulated data_trscr, data_cnv, data_meth, gene_list, pf_meth and exp_grp.
dmprocr_study = generate_fakestudy()

### perform differential analysis of data_trscr data
### fitType is set to "mean" because the simulated dataset is small (i.e. 500 genes)
dmprocr_result = RNAseq_diffAnalysis(data_trscr = dmprocr_study$data_trscr, 
                                     exp_grp = dmprocr_study$exp_grp, 
                                     gene_list = dmprocr_study$gene_list, 
                                     alpha = 0.05, 
                                     contrast = c("sample","case","ctrl"),
                                     fitType="mean")

### perform differential analysis of data_trscr data only of female patient
### fitType is set to "mean" because the simulated dataset is small (i.e. 500 genes)
filter_indiv = rownames(dmprocr_study$exp_grp[dmprocr_study$exp_grp$gender == "F", ])
dmprocr_result_f = RNAseq_diffAnalysis(data_trscr = dmprocr_study$data_trscr, 
                                     exp_grp = dmprocr_study$exp_grp, 
                                     gene_list = dmprocr_study$gene_list,
                                     filter_indiv = filter_indiv,
                                     alpha = 0.05, 
                                     contrast = c("sample","case","ctrl"),
                                     fitType="mean")

### perform linear regression on data_trscr and data_cnv  
### (be careful to use normalized count data as input)
dmprocr_lin_reg = RNAseq_cnv_reg(data_ntrscr = dmprocr_result$data_ntrscr, 
                                     data_cnv = dmprocr_study$data_cnv,
                                     exp_grp = dmprocr_study$exp_grp, 
                                     gene_list = dmprocr_result$result_list, 
                                     contrast = c("sample","case","ctrl"))

### you can change the filter on CNV distribution 
### and look only at a selected subset of indivuals (here females)
filter_indiv = rownames(dmprocr_study$exp_grp[dmprocr_study$exp_grp$gender == "F", ])
dmprocr_lin_reg_f =  RNAseq_cnv_reg(data_ntrscr = dmprocr_result$data_ntrscr, 
                                 data_cnv = dmprocr_study$data_cnv,
                                 exp_grp = dmprocr_study$exp_grp, 
                                 gene_list = dmprocr_result$result_list,
                                 filter_indiv = filter_indiv,
                                 contrast = c("sample","case","ctrl"),
                                 cnv_filter = c(0.05, 0.95))

### perform DE analysis on individuals with no cnv variation 
### (be careful to use normalized count data as input)
dmprocr_lin_reg_no_cnv = noCNV_diffAnalysis(data_ntrscr = dmprocr_result$data_ntrscr, 
                                     data_cnv = dmprocr_study$data_cnv,
                                     exp_grp = dmprocr_study$exp_grp, 
                                     gene_list = dmprocr_lin_reg, 
                                     contrast = c("sample","case","ctrl"))

### you can change the filter on CNV 
### (considere no cnv variation of cnv values between -0.3 and 0.3) 
### and look only at a selected subset of indivuals (here females)
filter_indiv = rownames(dmprocr_study$exp_grp[dmprocr_study$exp_grp$gender == "F", ])
dmprocr_lin_reg_no_cnv_f =  noCNV_diffAnalysis(data_ntrscr = dmprocr_result$data_ntrscr, 
                                 data_cnv = dmprocr_study$data_cnv,
                                 exp_grp = dmprocr_study$exp_grp, 
                                 gene_list =  dmprocr_lin_reg_f,
                                 filter_indiv = filter_indiv,
                                 contrast = c("sample","case","ctrl"),
                                 no_cnv_filter = c(-0.3, 0.3))

### Select candidates based on thresholds for padj, z_score and noCNVlog2FC 
candidates = select_candidates(gene_list = dmprocr_lin_reg_no_cnv,
                               t_padj = 0.001, 
                               t_z_score = 0.3, 
                               t_noCNVlog2FC = 2)



}

