% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translocateProfile.R
\name{translocateProfile}
\alias{translocateProfile}
\title{translocateProfile}
\usage{
translocateProfile(m1, bslide, bwin)
}
\arguments{
\item{m1}{is a dmProfile dataframe to translocate using bslide}

\item{bslide}{is the slide value divided by the width of each bin in dmProfile (= by.interp)}

\item{bwin}{is the window parameter divided by the width of each bin in dmProfile (= by.interp)}
}
\description{
Produce a list of length bslide + 1 where each element is a dmProfile dataframe translocated around the tss
}
