% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trscr_meth_analysis.R
\name{trscr_meth_analysis}
\alias{trscr_meth_analysis}
\title{Analyse transcriptome and methylome study}
\usage{
trscr_meth_analysis(gene_symbol, s_cnv, s_meth, s_trscr, gene_indexed_probes)
}
\arguments{
\item{gene_symbol}{the symbol of the gene}

\item{s_cnv}{cnv study}

\item{s_meth}{methylome study}

\item{s_trscr}{transcriptome study}

\item{gene_indexed_probes}{a list of probes, indexed by gene_symbol}
}
\description{
This function analyse transcriptome and methylome study
}

