% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustdmProfil.R
\name{clustdmProfile}
\alias{clustdmProfile}
\title{clustdmProfile}
\usage{
clustdmProfile(mat, fill_NA = TRUE, labels)
}
\arguments{
\item{mat}{a distance Matrix}

\item{fill_NA}{boolean specifying if NA should be imputed, need to be kept TRUE for now if there are NA in the matrix}

\item{labels}{A vector of genes names, can be extracted from a list of dmProfile with getProfileId}
}
\value{
A list of 2 object, the hclust result and a list containing as much vectors of geneNames as the number of cluster selected.
}
\description{
Plot a dendrogram from a distance matrix, select cluster by clicks on the graphical window, click finish to name of gene in each cluster selected
}
