% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmProfile.R
\name{compute_gene_meth_profile}
\alias{compute_gene_meth_profile}
\title{compute_gene_meth_profile}
\usage{
compute_gene_meth_profile(gene, meth_data, meth_platform, pf_pos_colname,
  pf_chr_colname, win, slide, interp.by, apply_func = apply)
}
\arguments{
\item{gene}{A list that describe gene in bed format.}

\item{meth_data}{A matrix of row methylation data with TCGA sample ids as column names and probes ids as row names.}

\item{meth_platform}{a data.frame with metadata types as columns names and probes ids as row names.}

\item{pf_pos_colname}{string matching the name of the column in the platform that contain the position information of probes}

\item{pf_chr_colname}{string matching the name of the column in the platform that contain the chromosome information of probes}

\item{win}{is the width of the window on the chromosome in bp where the function will fetch probes position and differential methylation value}

\item{slide}{is the maximum width slide you'll alow when comparing two curve}

\item{interp.by}{is resolution at which the function interpolate the probes signal}

\item{apply_func}{Function that will be used for apply.}
}
\description{
Generate  differential methylation profile for a gene around the transcription start side
}

