% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmDistance.R
\name{dmDistance}
\alias{dmDistance}
\title{dmDistance}
\usage{
dmDistance(dmprofileList, frechet = FALSE)
}
\arguments{
\item{dmprofileList}{a list of dmProfile}

\item{frechet}{a boolean specify if frechet distance will be computed.}
}
\description{
Perform either a frechet distance from the kmlShape package or the eucliddean distance modified from this package (default) between two dmProfile. Return a distance matrix.
[warning]Carefully use the frechet distance as it can be heavy computing when dealing with large set of profile. Complexity of the profile also weight on the memory usage.
}
\examples{
### generate random samples ids
randomDataset          <- dmRandomDataset()
### get a methylation differential table
meth_study          <- randomDataset$meth_study
tumoral_ref         <- rownames(meth_study$exp_grp)[meth_study$exp_grp$ref=="case"]
control_ref         <- rownames(meth_study$exp_grp)[meth_study$exp_grp$ref=="ctrl"]
diff_meth_study     <- dmTable(meth_study$data,
                              meth_study$platform,
                              meth_study$exp_grp,
                              tumoral_ref,
                              control_ref)
### get dmProfiles for multiple genes
genes        <- randomDataset$genes 
alldmProfile <- getalldmProfile(diff_meth_study,
                                bedfile = genes, slide = 500,
                                pf_chr_colname = "chr",
                                pf_pos_colname = "pos")

}
