% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_analysis.R
\name{dmDistance}
\alias{dmDistance}
\title{dmDistance}
\usage{
dmDistance(profiles, frechet = FALSE)
}
\arguments{
\item{profiles}{a list of dmProfile}

\item{frechet}{a boolean specify if frechet distance will be computed.}
}
\description{
Perform either a frechet distance from the kmlShape package or the eucliddean distance modified from this package (default) between two dmProfile. Return a distance matrix.
[warning]Carefully use the frechet distance as it can be heavy computing when dealing with large set of profile. Complexity of the profile also weight on the memory usage.
}
