% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmTable.R
\name{dmTable}
\alias{dmTable}
\title{dmTable}
\usage{
dmTable(data, platform, exp_group, TumoralRef, ControlRef, NARM = TRUE)
}
\arguments{
\item{data}{A matrix of row methylation data with TCGA sample ids as column names and probes ids as row names.}

\item{platform}{a dataframe with metadata types as columns names and probes ids as row names.}

\item{exp_group}{a dataframe of metadata about the samples. Rows are ids and columns information type.}

\item{TumoralRef}{a vector of ids corresponding to tumoral samples.}

\item{ControlRef}{a vector of ids corresponding to control samples.}

\item{NARM}{will delete empty rows in data and platform if TRUE.}
}
\description{
Generate differential methylation data table from beta values. It basically extract all tumorous samples and controls. Then it computes the difference between each tumorous and the mean of control.
}
\examples{
###generate random samples ids
sampleId <- paste0(rep("sample", 10), 1:10)
idProbes <- paste0(rep("probe", 10), 1:10)

###generate data
data <- matrix(runif(100,0,1), 10, 10)
colnames(data) <- sampleId
rownames(data) <- idProbes

###generate exp_group
exp_group <- as.data.frame(matrix(rep("x", 40),  10, 4), row.names = sampleId)
exp_group$ref <- rep(c("01","11"),5)

###generate platfrom
platform  <-as.data.frame(matrix(rep("x", 50),  10, 5), row.names = idProbes)

###get indx of ref
TumoralRef <- rownames(exp_group[exp_group$ref == "01", ])
ControlRef <- rownames(exp_group[exp_group$ref == "11", ])

###get a methylation differential table
diff_table <- dmTable(data, platform, exp_group, ControlRef = ControlRef, TumoralRef = TumoralRef)
}
