% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmTable.R
\name{dmTable}
\alias{dmTable}
\title{dmTable}
\usage{
dmTable(data, platform, exp_grp, tumoral_ref, control_ref, RM_NA_ROWS = TRUE)
}
\arguments{
\item{data}{A matrix of row methylation data with TCGA sample ids as column names and probes ids as row names.}

\item{platform}{a dataframe with metadata types as columns names and probes ids as row names.}

\item{exp_grp}{a dataframe of metadata about the samples. Rows are ids and columns information type.}

\item{tumoral_ref}{a vector of ids corresponding to tumoral samples.}

\item{control_ref}{a vector of ids corresponding to control samples.}

\item{RM_NA_ROWS}{will delete empty rows in data and platform if TRUE.}
}
\description{
Generate differential methylation data table from beta values. It basically extract all tumorous samples and controls. Then it computes the difference between each tumorous and the mean of control.
}
\examples{
### generate random samples ids
randomDataset          <- dmRandomDataset()
### get a methylation differential table
meth_study          <- randomDataset$meth_study
tumoral_ref         <- rownames(meth_study$exp_grp)[meth_study$exp_grp$ref=="case"]
control_ref         <- rownames(meth_study$exp_grp)[meth_study$exp_grp$ref=="ctrl"]
diff_meth_study     <- dmTable(meth_study$data,
                              meth_study$platform,
                              meth_study$exp_grp,
                              tumoral_ref,
                              control_ref)
}
