% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmProfile.R
\name{get_probe_names}
\alias{get_probe_names}
\title{A Function That Extracts Probe Names of a Corresponding Gene from Platform Data}
\usage{
get_probe_names(gene, meth_platform, pf_chr_colname = "Chromosome",
  pf_pos_colname = "Start", up_str = 5000, dwn_str = 5000)
}
\arguments{
\item{gene}{A vector describing the gene (line of a bed file).}

\item{meth_platform}{A data frame describing CpG positions.}

\item{pf_pos_colname}{string matching the name of the column in the platform that contain the chromosome on which we find a probes.}

\item{up_str}{An integer specifying up stream size (in bp).}

\item{dwn_str}{An integer specifying down stream size (in bp).}

\item{pf_pos_colname}{string matching the name of the column in the platform that contain the position information of probes.}
}
\value{
A vector of probe names
}
\description{
A Function That Extracts Probe Names of a Corresponding Gene from Platform Data
}

