% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDifferentialTable.R
\name{getDifferentialTable}
\alias{getDifferentialTable}
\title{getDifferentialTable}
\usage{
getDifferentialTable(data, platform, exp_group)
}
\arguments{
\item{data}{A matrix of row methylation data with TCGA sample ids as column names and probes ids as row names}

\item{platform}{a dataframe with metadata types as columns names and probes ids as row names}

\item{exp_group}{a dataframe of metadata about the samples. Rows are ids and columns information type.}
}
\description{
This function generate a differential methylation data table from raw data. It basically extract all tumorous samples and controls. Then it computes the difference between each tumorous and the mean of control
}
