% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethylInfo.R
\name{getMethylInfo}
\alias{getMethylInfo}
\title{getMethylInfo}
\usage{
getMethylInfo(diff_meth_study, bedline, win = 5000,
  pf_chr_colname = "Chromosome", pf_pos_colname = "Start")
}
\arguments{
\item{diff_meth_study}{list composed of a methylation differential table, an exp_grp dataframe and a platform}

\item{bedline}{is a line of of the bedfile dataframe}

\item{win}{is the width of the window on the chromosome in bp where the function will fetch probes position and differential methylation value. Default is 5000 bp}

\item{pf_chr_colname}{string matching the name of the column in the platform that contain the chromosome on which we find a probes}

\item{pf_pos_colname}{string matching the name of the column in the platform that contain the position information of probes}
}
\description{
Return a subset of diff_meth_study list for a given gene bedline and a size of window
}
\examples{
### generate random samples ids
randomDataset          <- dmRandomDataset()
### get a methylation differential table
meth_study          <- randomDataset$meth_study
tumoral_ref         <- rownames(meth_study$exp_grp)[meth_study$exp_grp$ref=="case"]
control_ref         <- rownames(meth_study$exp_grp)[meth_study$exp_grp$ref=="ctrl"]
diff_meth_study     <- dmTable(meth_study$data,
                              meth_study$platform,
                              meth_study$exp_grp,
                              tumoral_ref,
                              control_ref)
### get the gene probes information
genes = meth_study$genes[2,]
pf_chr_colname  <- "chr"
pf_pos_colname  <- "pos"
gene_study <- getMethylInfo(diff_meth_study,
                            bedline = genes,
                            pf_chr_colname = "chr",
                            pf_pos_colname = "pos")
}
