% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_analysis.R
\name{interpolate_gene}
\alias{interpolate_gene}
\title{interpolate_gene}
\usage{
interpolate_gene(vec, meth_probe_pos, xf, tss, updwn_str, slide)
}
\arguments{
\item{vec}{A numeric vector specifying differential methylation signal.}

\item{meth_probe_pos}{is a vector of probes position on the chromosome.}

\item{xf}{coordinates of interpolation}

\item{tss}{the transcription start site on the same chromosome.}

\item{updwn_str}{is the width of the window on the chromosome in bp where the function will fetch probes position and differential methylation value, default is 5000.}

\item{slide}{is the maximum width slide you'll alow when comparing two curve, default is 0.}
}
\description{
Build a interpolated signal of differential methylation value for a gene.
}

