% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trscr_analysis.R
\name{noCNV_diffAnalysis}
\alias{noCNV_diffAnalysis}
\title{Calculate differential expression on a subset of patient.}
\usage{
noCNV_diffAnalysis(data_ntrscr, data_cnv, exp_grp, gene_list,
  filter_indiv = "no_filter", contrast = c("tissue_status", "patho",
  "normal"), no_cnv_filter = c(-0.2, 0.2))
}
\arguments{
\item{data_ntrscr}{A \code{data} matrix that contains normalized RNAseq counts (from DESeq2 analysis). Columns correspond to indivuals, row correspond to genes.}

\item{data_cnv}{A \code{data} matrix that contains CNV data}

\item{exp_grp}{A \code{exp_grp} dataframe that contains metadatas on \code{data_trscr} individuals.}

\item{gene_list}{A \code{gene_list} bedfile containing the genes to screen for differential expression.}

\item{filter_indiv}{A vector of individual names to be screened for differential expression. Optionnal (set on "no_filter" by default).}

\item{contrast}{A vector containing the constrast to be used to estimate the logarithmic fols change. By default: c("tissue_status","patho","normal")}

\item{no_cnv_filter}{A vector of two values indicating the CNV values threshold for no CNV, by default set to c(-0.2, 0.2).}
}
\value{
A \code{gene_list} table including a noCNVlog2FC value associated with differential expression between sample without CNV.
}
\description{
Calculate differential expression on a subset of patient, in particular those with no CNV
}

