% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meth_analysis.R
\name{plotdmProfile}
\alias{plotdmProfile}
\title{plotdmProfile}
\usage{
plotdmProfile(meth_profiles, i = 1)
}
\arguments{
\item{meth_profiles}{a list of dmProfile}

\item{i}{integer specifying the i-st profile to plot}
}
\description{
Produce a list of differential methylation profile plot with ggplot2.
}

