% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffAnalysis.R
\name{RNAseq_diffAnalysis}
\alias{RNAseq_diffAnalysis}
\title{Perform differential analysis of RNAseq data.}
\usage{
RNAseq_diffAnalysis(data_trscr, exp_grp, gene_list,
  filter_indiv = "no_filter", alpha = 0.05, contrast = c("sample", "01",
  "11"), fitType = "parametric")
}
\arguments{
\item{data_trscr}{A \code{data} matrix that contains transcriptome information (RNAseq counts from HTseq). Columns correspond to indivuals, row correspond to genes.}

\item{exp_grp}{A \code{exp_grp} dataframe that contains metadatas on \code{data_trscr} individuals.}

\item{gene_list}{A \code{gene_list} bedfile containing the genes to screen for differential expression.}

\item{filter_indiv}{A vector of individual names to be screened for differential expression. Optionnal (set on "no_filter" by default).}

\item{alpha}{A parameter to indicate the significance cutoff used by \code{DESeq2::results} funtion for optimizing the independent filtering (by default 0.05). If the adjusted p-value cutoff (FDR) will be a value other than 0.1, alpha should be set to that value.}

\item{contrast}{A vector containing the constrast to be used to estimate the logarithmic fols change}

\item{fitType}{A DESeq fitting paramater, by default set to "parametric"}
}
\value{
A \code{gene_list} table including log2FoldChange and adjusted p-value (padj) computed by DESeq2 and a \code{data_ntrscr} matrix of normalized counts.
}
\description{
Perform differential analysis of transcriptomic data (RNAseq) using DESeq2 R package.
}
\examples{
### generate a list of simulated data_trscr, data_cnv, data_meth, gene_list, pf_meth and exp_grp.
dmprocr_study = generate_fakestudy()

### perform differential analysis of data_trscr data
### fitType is set to "mean" because the simulated dataset is small (i.e. 500 genes)
dmprocr_result = RNAseq_diffAnalysis(data_trscr = dmprocr_study$data_trscr, 
                                     exp_grp = dmprocr_study$exp_grp, 
                                     gene_list = dmprocr_study$gene_list, 
                                     alpha = 0.05, 
                                     contrast = c("sample","case","ctrl"),
                                     fitType="mean")

### perform differential analysis of data_trscr data only of female patient
### fitType is set to "mean" because the simulated dataset is small (i.e. 500 genes)
filter_indiv = rownames(dmprocr_study$exp_grp[dmprocr_study$exp_grp$gender == "F", ])
dmprocr_result_f = RNAseq_diffAnalysis(data_trscr = dmprocr_study$data_trscr, 
                                     exp_grp = dmprocr_study$exp_grp, 
                                     gene_list = dmprocr_study$gene_list,
                                     filter_indiv = filter_indiv,
                                     alpha = 0.05, 
                                     contrast = c("sample","case","ctrl"),
                                     fitType="mean")

}
