% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trscr_analysis.R
\name{select_candidates}
\alias{select_candidates}
\title{Generate a candidate table}
\usage{
select_candidates(gene_list, padj_thresh = 0, z_score_thresh,
  noCNVlog2FC_thresh)
}
\arguments{
\item{gene_list}{A \code{gene_list} bedfile containing the log2FC, zscore and noCNVlog2FC values.}

\item{padj_thresh}{A threshold p-value under which expression is significantly different by DESeq analysis}

\item{z_score_thresh}{A threshold zscore value, candidate genes should display a zcore value such as -t_zscore<zscore<t_zscore}

\item{noCNVlog2FC_thresh}{A threshold noCNVlog2FC, candidate genes should display a noCNVlog2FC value such as noCNVlog2FC > noCNVlog2FC_thresh or noCNVlog2FC < noCNVlog2FC_thresh}
}
\value{
A \code{gene_list} table including of candidate genes
}
\description{
Generate a candidate table
}
